+++ packages/datadog_flutter_plugin/lib/datadog_flutter_plugin.dart
@@ -3,7 +3,6 @@
 // Copyright 2016-Present Datadog, Inc.
 
 import 'dart:async';
-import 'dart:io';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
@@ -11,8 +10,8 @@ import 'package:meta/meta.dart';
 
 import 'src/attributes.dart';
 import 'src/datadog_configuration.dart';
+import 'src/datadog_plugin.dart';
 import 'src/datadog_sdk_platform_interface.dart';
-import 'src/datadog_tracking_http_client.dart';
 import 'src/helpers.dart';
 import 'src/internal_logger.dart';
 import 'src/logs/ddlogs.dart';
@@ -23,8 +22,15 @@ import 'src/version.dart' show ddPackageVersion;
 
 export 'src/attributes.dart' show DatadogConfigKey;
 export 'src/datadog_configuration.dart';
+export 'src/datadog_plugin.dart';
 export 'src/rum/ddrum.dart'
-    show RumHttpMethod, RumUserActionType, RumErrorSource, RumResourceType;
+    show
+        RumHttpMethod,
+        RumUserActionType,
+        RumErrorSource,
+        RumResourceType,
+        rumMethodFromMethodString,
+        resourceTypeFromContentType;
 export 'src/rum/navigation_observer.dart'
     show
         DatadogNavigationObserver,
@@ -66,6 +72,8 @@ class DatadogSdk {
   List<String> _firstPartyHosts = [];
   RegExp? _firstPartyRegex;
 
+  final Map<Type, DatadogPlugin> _plugins = {};
+
   /// A list of first party hosts for tracing. Note that this is an unmodifiable
   /// list. If you need to add a host, call the setter for [firstPartyHosts]
   List<String> get firstPartyHosts => List.unmodifiable(_firstPartyHosts);
@@ -95,6 +103,10 @@ class DatadogSdk {
     internalLogger.sdkVerbosity = value;
     unawaited(_platform.setSdkVerbosity(value));
   }
+
+  /// Get an instance of a DatadogPlugin that was registered with
+  /// [DdSdkConfiguration.addPlugin]
+  T? getPlugin<T>() => _plugins[T.runtimeType] as T?;
 
   /// This function is not part of the public interface for Datadog, and may not
   /// be available in all targets. Used for integration and E2E testing purposes only.
@@ -142,10 +154,6 @@ class DatadogSdk {
 
     await _platform.initialize(configuration, logCallback: _platformLog);
 
-    if (configuration.trackHttpClient) {
-      HttpOverrides.global = DatadogTrackingHttpOverrides(DatadogSdk.instance);
-    }
-
     if (configuration.loggingConfiguration != null) {
       _logs = createLogger(configuration.loggingConfiguration!);
     }
@@ -154,6 +162,17 @@ class DatadogSdk {
     }
     if (configuration.rumConfiguration != null) {
       _rum = DdRum(internalLogger);
+    }
+
+    for (final pluginConfig in configuration.additionalPlugins) {
+      var plugin = pluginConfig.create(this);
+      if (_plugins.containsKey(plugin.runtimeType)) {
+        internalLogger.error(
+            'Attempting to setup two plugins of the same type: ${plugin.runtimeType}. The second plugin will be ignored.');
+      } else {
+        plugin.initialize();
+        _plugins[plugin.runtimeType] = plugin;
+      }
     }
   }